+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * [#815](https://github.com/bbatsov/rubocop/issues/815): Fix a false positive for heredocs with blank lines in them in `EmptyLines`. ([@bbatsov][])
 * Auto-correction is now more robust and less likely to die because of `RangeError` or "clobbering". ([@jonas054][])
 * Offenses always reported in order of position in file, also during `--auto-correct` runs. ([@jonas054][])
+* Fix problem with `[Corrected]` tag sometimes missing in output from `--auto-correct` runs. ([@jonas054][])
 
 ## 0.18.1 (02/02/2014)